+++ kamehouse-groot/public/kame-house-groot/api/v1/admin/kamehouse-shell/execute.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * [EXTERNAL] Endpoint: /kame-house-groot/api/v1/admin/kamehouse-shell/execute.php (GET)
+ * Endpoint: /kame-house-groot/api/v1/admin/kamehouse-shell/execute.php (GET)
  * 
  * Url parameters: 
  *  - script: Script to execute with relative path to ${HOME}/programs/kamehouse-shell/bin
@@ -8,22 +8,22 @@
  * 
  * @author nbrest
  */
-$kameHouseShellExecutor = new KameHouseShellExecutor();
-$kameHouseShellExecutor->execute();
+$executeApi = new ExecuteApi();
+$executeApi->main();
 
-class KameHouseShellExecutor {
+class ExecuteApi {
 
   /**
    * Execute the kamehouse shell script.
    */
-  public function execute() {
-    global $kameHouse;
+  public function main() {
     $documentRoot = realpath($_SERVER["DOCUMENT_ROOT"]);
-    require_once("$documentRoot/kame-house-groot/api/v1/commons/kamehouse.php");
-    require_once("$documentRoot/kame-house-groot/api/v1/admin/kamehouse-shell/kamehouse-shell.php");
-    require_once("$documentRoot/kame-house-groot/api/v1/auth/kamehouse-auth.php");
+    require_once("$documentRoot/kame-house-groot/api/v1/kamehouse/commons/kamehouse.php");
+    require_once("$documentRoot/kame-house-groot/api/v1/kamehouse/admin/kamehouse-shell/kamehouse-shell.php");
+    require_once("$documentRoot/kame-house-groot/api/v1/kamehouse/auth/kamehouse-auth.php");
     $kameHouse->auth->authorizeApi();
     $kameHouse->shell->execute();
   }
+  
 }
 ?>