+++ kamehouse-groot/public/kame-house-groot/api/v1/commons/session/status.php
@@ -1,27 +1,27 @@
 <?php 
 /**
- * [EXTERNAL] Endpoint: /kame-house-groot/api/v1/commons/session/status.php (GET)
+ * Endpoint: /kame-house-groot/api/v1/commons/session/status.php (GET)
  * 
  * Get the current session status.
  * 
  * @author nbrest
  */
-$kameHouseSessionStatus = new KameHouseSessionStatus();
-$kameHouseSessionStatus->load();
+$statusApi = new StatusApi();
+$statusApi->main();
 
-class KameHouseSessionStatus {
+class StatusApi {
 
   /**
    * Load the session status.
    */
-  public function load() {
-    global $kameHouse;
+  public function main() {
     $documentRoot = realpath($_SERVER["DOCUMENT_ROOT"]);
-    require_once("$documentRoot/kame-house-groot/api/v1/commons/kamehouse.php");
-    require_once("$documentRoot/kame-house-groot/api/v1/admin/kamehouse-shell/kamehouse-shell.php");
-    require_once("$documentRoot/kame-house-groot/api/v1/auth/kamehouse-auth.php");
-    require_once("$documentRoot/kame-house-groot/api/v1/commons/session/kamehouse-session.php");
+    require_once("$documentRoot/kame-house-groot/api/v1/kamehouse/commons/kamehouse.php");
+    require_once("$documentRoot/kame-house-groot/api/v1/kamehouse/admin/kamehouse-shell/kamehouse-shell.php");
+    require_once("$documentRoot/kame-house-groot/api/v1/kamehouse/auth/kamehouse-auth.php");
+    require_once("$documentRoot/kame-house-groot/api/v1/kamehouse/commons/session/kamehouse-session.php");
     $kameHouse->session->getStatus();
   }
+  
 }
 ?>