+++ kamehouse-groot/public/kame-house-groot/api/v1/kamehouse/admin/kamehouse-shell/kamehouse-shell.php
@@ -1,9 +1,5 @@
 <?php
 /**
- * [INTERNAL] Endpoint: /kame-house-groot/api/v1/admin/kamehouse-shell/kamehouse-shell.php (GET)
- * 
- * To be imported from other php files. Not to be directly called from frontend code.
- * 
  * Class definition of kamehouse shell.
  * 
  * @author nbrest