+++ kamehouse-groot/public/kame-house-groot/api/v1/test/test.php
@@ -1,67 +1,27 @@
 <?php 
 /**
- * [EXTERNAL] Endpoint: /kame-house-groot/api/v1/test/test.php (GET)
+ * Endpoint: /kame-house-groot/api/v1/test/test.php (GET)
  * 
- * Test php functionality
+ * Test php functionality.
  * 
  * @author nbrest
  */
-$kameHouseTest = new KameHouseTest();
-$kameHouseTest->runAll();
+$testApi = new TestApi();
+$testApi->main();
 
-class KameHouseTest {
-
-  /**
-   * Run all test functions.
-   */
-  public function runAll() {
-    global $kameHouse;
-    $this->init();
-  
-    $kameHouse->logger->info("Accessing test page");
-    // print server info. ***** DON'T LEAVE THIS UNCOMMENTED *****
-    // Gives a lot of info of the server
-    echo "<h1>print server info</h1>";
-    //print_r($_SERVER);
-    //print_r ($_SESSION);
-
-    // print all headers
-    echo "<h1>print all headers</h1>";
-    foreach (getallheaders() as $name => $value) {
-      echo "$name: $value<br>";
-    }
-  
-    // Test function defined in the same test.php file
-    $this->testFunction();
-  
-    $this->runExamples();
-    
-    // print phpinfo. ***** DON'T LEAVE THIS UNCOMMENTED *****
-    //phpinfo();
-  }
+class TestApi {
 
   /**
-   * Init test.
+   * Run all test and example functions.
    */
-  private function init() {
+  public function main() {
     $documentRoot = realpath($_SERVER["DOCUMENT_ROOT"]);
-    require_once("$documentRoot/kame-house-groot/api/v1/commons/kamehouse.php");
-  }
-  
-  /** 
-   * I can define functions here as well and call them in the API endpoint 
-   */
-  private function testFunction() {
-    echo "<br><br>testFunction: mada mada dane:<br><br>";
+    require_once("$documentRoot/kame-house-groot/api/v1/kamehouse/commons/kamehouse.php");
+    require_once("$documentRoot/kame-house-groot/api/v1/kamehouse/test/test.php");
+    require_once("$documentRoot/kame-house-groot/api/v1/kamehouse/commons/examples/examples.php");
+    $kameHouseTest->runAll();
+    $kameHouseExamples->runAll();
   }
 
-  /**
-   * Run kamehouse examples.
-   */
-  private function runExamples() {
-    $documentRoot = realpath($_SERVER["DOCUMENT_ROOT"]);
-    require_once("$documentRoot/kame-house-groot/api/v1/commons/examples/examples.php");
-  }
-
-} // KameHouseTest
+}
 ?>